^R Extract GitHub-release helpers into internal/host/release (go build/test green; first sub-PR of the internal/hostutil split, no behavior change)
internal/hostutil mixes nine unrelated concerns (sessions, GitHub
releases, paths, Colima, locks, audit, mounts, random hex, launcher).
This is the first incremental split per /sethify decision D7 — the
GitHub-release helpers move into a focused internal/host/release
package, leaving the rest of hostutil intact for follow-up PRs.

Moved:
  - releaseCreatePayload / releaseAsset / releaseResponse /
    releaseBundleManifest types
  - WriteGitHubReleaseCreatePayload
  - WriteGitHubReleaseMetadata
  - EncodeReleaseAssetName
  - WriteReleaseBundleManifest
  - the four matching tests
  → internal/host/release/release.go + release_test.go

Updated cmd/workcell-hostutil/main.go to import the new package; the
runRelease dispatcher still routes the same four subcommands
(create-payload, metadata, encode-name, bundle-manifest) without any
runtime contract change.

internal/hostutil/hostutil.go shrinks from 166 to 52 lines (only path
canonicalization and the expandUser helper remain). The test file
drops the four release tests; nothing else moves yet.

Net: 275 insertions, 234 deletions across 5 files. Subsequent PRs in
this series will extract sessions, launcher, audit-path, and other
clusters.

Relates to /sethify Run-1 Code Quality 🟡 (hostutil mixed concerns).
